Add more functions to UuidUtil.h
