+++ include/RobotRaconteurCompanion/Util/UuidUtil.h
@@ -24,6 +24,7 @@
 #include "com__robotraconteur__uuid.h"
 #include <boost/uuid/uuid.hpp>
 #include <boost/uuid/uuid_io.hpp>
+#include "UtilMacros.h"
 
 #pragma once
 
@@ -34,31 +35,53 @@ namespace Companion
 namespace Util
 {
     /**
-     * @brief Parse a UUID from a string
+     * Parse a UUID from a string.
      * 
-     * @param string_uuid The string to parse
-     * @return com::robotraconteur::uuid::UUID The parsed UUID
+     * Parses a string representation of a UUID and returns a com::robotraconteur::uuid::UUID object.
+     * 
+     * @param string_uuid The string to parse.
+     * @return The parsed UUID.
      */
-    static com::robotraconteur::uuid::UUID ParseUuid(const std::string& string_uuid)
-    {
-        com::robotraconteur::uuid::UUID ret;
-        boost::uuids::uuid boost_uuid = boost::lexical_cast<boost::uuids::uuid>(string_uuid);
-        std::copy(boost_uuid.begin(), boost_uuid.end(), ret.a.begin());
-        return ret;
-    }
+    ROBOTRACONTEUR_COMPANION_UTIL_API com::robotraconteur::uuid::UUID ParseUuid(const std::string& string_uuid);
 
     /**
-     * @brief Convert a UUID to a string
+     * Convert a UUID to a string.
      * 
-     * @param uuid The UUID to convert
-     * @return std::string The string
+     * Converts a com::robotraconteur::uuid::UUID object into a string representation.
+     * 
+     * @param uuid The UUID to convert.
+     * @return The string representation of the UUID.
      */
-    static std::string UuidToString(com::robotraconteur::uuid::UUID uuid)
-    {
-        boost::uuids::uuid boost_uuid;
-        std::copy(uuid.a.begin(),uuid.a.end(),boost_uuid.begin());
-        return boost::lexical_cast<std::string>(boost_uuid);
-    }
+    ROBOTRACONTEUR_COMPANION_UTIL_API std::string UuidToString(com::robotraconteur::uuid::UUID uuid);
+
+    /**
+     * Convert a boost::uuids::uuid to com::robotraconteur::uuid::UUID.
+     * 
+     * Converts a boost::uuids::uuid object to a com::robotraconteur::uuid::UUID object.
+     * 
+     * @param boost_uuid The boost::uuids::uuid to convert.
+     * @return The converted UUID.
+     */
+    ROBOTRACONTEUR_COMPANION_UTIL_API com::robotraconteur::uuid::UUID BoostUuidToUuid(const boost::uuids::uuid& boost_uuid);
+
+    /**
+     * Convert a com::robotraconteur::uuid::UUID to boost::uuids::uuid.
+     * 
+     * Converts a com::robotraconteur::uuid::UUID object to a boost::uuids::uuid object.
+     * 
+     * @param uuid The com::robotraconteur::uuid::UUID to convert.
+     * @return The converted boost::uuids::uuid.
+     */
+    ROBOTRACONTEUR_COMPANION_UTIL_API boost::uuids::uuid UuidToBoostUuid(const com::robotraconteur::uuid::UUID& uuid);
+
+    /**
+     * @brief Generate a new random UUID
+     * 
+     * Generates a new random UUID using boost::uuids::random_generator().
+     * 
+     * @return com::robotraconteur::uuid::UUID The generated random UUID
+     */
+    ROBOTRACONTEUR_COMPANION_UTIL_API com::robotraconteur::uuid::UUID NewRandomUuid();
 }
 }
 }
